Add support for extension modules
Extension modules are jar files, and can be loaded like normal modules.
This means, we get back a class, which can be instantiated.
The methods on the class are defined by the @Primitive annotations on node classes.
These primitives have to be made available via the list of factories to the primitive loader.

- added basic tests
- added test extension
- added extension primitive loader
- some refactoring of mixin builder and parser
- added Eclipse and Ant setup
- added documentation

Added extension test Eclipse and Ant setup

Signed-off-by: Stefan Marr <git@stefan-marr.de>