Use CachedSerializationNode in MessageSerializationNode
Since classes are merely id, but not the structure, it’s less stable to specialize on classes. Each new object with inner classes will generate new class objects for the inner classes. Thus, the specialization would need to adapt. Therefore, avoid using the argumentClasses in MSN. I am now using the CachedSerializationNode, which also still needs to be adapted though.

Signed-off-by: Stefan Marr <git@stefan-marr.de>