Object safepoint operations need to be in finally blocks consistently
This is a correctness precaution.
All subsequent register/unregister operations following an initial unregister/register need to be in a finally clause if the operations in-between could cause any kind of exception, which most can.

For consistency, it should be used really everywhere.

There are only two points where it might be acceptable, but should be still fixed: init of object system, and in the executeApplication method.

Signed-off-by: Stefan Marr <git@stefan-marr.de>