Optimize the ValueCheckNode
This node is used in the constructor of Value objects, which are not allowed to have any members that are mutable.

The now optimized node takes the layout of an object and gets all the slot accessors that can potentially access mutable objects, i.e., the slots containing non-primitives.
It constructs a tester object, which will then test these fields.

The fallback case, i.e., when we see too many layouts, is the old code, that simply goes over all slots and checks whether it contains a value.

Signed-off-by: Stefan Marr <git@stefan-marr.de>