Added a MultiPardardigmBench
This is based on the KomposBench.

The application is changed to record the sales for the fruits over time. The DataStorage actor, interfacing with a legacy system, uses transactions, to store the data.

The reporting actor is replaced by a forecast actor and uses a simple linear regression to forecast future sales. Each product is forecasted in parallel using f/j tasks, results are stored in an array, using a lock. It’s not really racy, but, well, it’s a demo…

Signed-off-by: Stefan Marr <git@stefan-marr.de>