[WIP] Use onInputValue for DynamicMetrics tool
This is an attempt to use the onInputValue method provided by Truffle’s instrumentation framework.

However, this doesn’t work directly because the way eager specialization works with using a node wrapping the primitive node.

Currently, there’s a Truffle wrapper node inserted between the eager node and the primitive node, which breaks the eager replacement system.

Options to solve this:
 1. avoid the use of eager primitive nodes by adding a fallback specialization that rewrites to a generic node to every primitive (or a suitable superclass)
 2. making the primitives not wrap-able (i.e., removing the @GenerateWrapperNode annotation from the super class it is currently on)

None of these changes seem trivial. So for the moment, I’ll not further pursue this.

Signed-off-by: Stefan Marr <git@stefan-marr.de>