Fix the profiling of block activations
Since we changed the block implementation to use the TruffleDSL, our stats where incomplete.

This fixes the implementation and collects the statistics by manually traversing the node structure to extract the counts from the CountingDirectCallNodes.

This is brittle and depends on the TruffleDSL not changing too much, but I suppose, there are no major changes expected.

Signed-off-by: Stefan Marr <git@stefan-marr.de>