Replace use of SourceCoordinate with int for startIndex
The start index is the only bit we actually use, so, no need to do more.
There are a few places that want line and column, so, we deal with those explicitly.

- track the startLastNonWhiteCharIdx now explicitly
- add lexer tests from TruffleSOM

The whole approach has been used in TruffleSOM since I started not using source sections there. Though, that’s too much of a change for the moment. This here is to fix bugs around source locations.

Signed-off-by: Stefan Marr <git@stefan-marr.de>